hack/mqtt/mqtt-example/requirements.txt
paho-mqtt == 1.5.0
python-qpid-proton == 0.30.0